# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Documentation for Cric Heros",
        "title": "Cric Heros API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "paths": {
...
        "/addPlayertoTeam": {
            "post": {
                "description": "Add player to team",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Team"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the team",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Array of players",
                        "name": "player",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/createTeam": {
            "post": {
                "description": "Creates a team",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Team"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the user",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Details of the team",
                        "name": "TeamDetails",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Team"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/decisionUpdate": {
            "put": {
                "description": "Updates the decison taken by the team after wining the toss",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Toss"
                ],
                "parameters": [
                    {
                        "description": "Descision Updated",
                        "name": "toss",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Toss"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/deleteTeamByID": {
            "delete": {
                "description": "Delete the team",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Team"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the team",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "ID of the user",
                        "name": "user_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/showTeamByID": {
            "post": {
                "description": "Shows the list of teams",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Team"
                ],
                "parameters": [
                    {
                        "description": "ID of the team",
                        "name": "team_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/showTeams": {
            "get": {
                "description": "Shows the list of teams",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Team"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the User",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/tossResult": {
            "post": {
                "description": "Give the random result of coin toss and which team won the toss",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Toss"
                ],
                "parameters": [
                    {
                        "description": "Toss Details",
                        "name": "toss",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Toss"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "models.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "status": {
                    "type": "string"
                }
            }
        },
        "models.Team": {
            "type": "object",
            "required": [
                "teamCaptain",
                "teamName",
                "teamType"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "playerId": {
                    "type": "string"
                },
                "teamCaptain": {
                    "type": "string"
                },
                "teamId": {
                    "type": "string"
                },
                "teamName": {
                    "type": "string"
                },
                "teamType": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "models.Toss": {
            "type": "object",
            "required": [
                "head_team",
                "match_id",
                "tail_team"
            ],
            "properties": {
                "decision": {
                    "type": "string"
                },
                "head_team": {
                    "type": "string"
                },
                "match_id": {
                    "type": "string"
                },
                "tail_team": {
                    "type": "string"
                },
                "toss_id": {
                    "type": "string"
                },
                "toss_won": {
                    "type": "string"
                }
            }
        }
    }
}